### 2025-04-07 #7910 @fabian-s
:wave: @cmparlettpelleriti @lucasmiranda42

would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
Please be so kind to reply even if you're not interested, so that I can start looking for alternative reviewers.
