### 2025-04-16 #8063 @editorialbot
[REVIEW]: medRCT: Causal mediation analysis estimating interventional effects mapped to a target trial in R
**Submitting author:** @t0ngchen (<a href="http://orcid.org/0000-0003-2751-2412">Tong Chen</a>)
**Repository:** https://github.com/T0ngChen/medRCT
**Branch with paper.md** (empty if default branch):
**Version:** v.0.1.1
