### 2025-05-11 T0ngChen/medRCT#23 @nhejazi
[joss] clarifying state-of-the-field and users
per criteria outlined in the ongoing review at https://github.com/openjournals/joss-reviews/issues/8063#issuecomment-2847173310, the software documentation should contain a statement of need, specifically

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

it seems that the package's `README` would be a good place to do this. while the interventional direct/indirect effects and target trial emulation are both mentioned in the `README` file and in the vignettes, the above criteria---explaining briefly *why* one would want to assess such effects via trial emulation (even if already explained in the cited methods paper)---are not yet satisfied. this may also provide an opportunity to explain why and where this package ought to be used relative to alternatives.
