### 2025-05-11 #8063 @nhejazi
i've nearly completed my review. overall, this looks like a nicely put together package, one that will be useful to applied scientists using this mediation analysis approach. a few points should be addressed prior to my completing the reviewer checklist above.

#### documentation -- statement of need

https://github.com/T0ngChen/medRCT/issues/23 details a few changes that could be made to help tick off the "statement of need"

### software paper

some statements in the software paper as currently made are either inaccurate or misleading.

1) please clarify what the "certain identifiability challenges associated with traditional estimands, known as the natural effects" are (also Pearl (2001) should be cited alongside Robins and Greenland (1992) for the NDE/NIE). explaining this in detail is beyond the scope of a software paper but this statement is too vague for the inexperienced reader to appreciate what exactly the challenges are.
2) the second half of the paragraph containing the excerpt in (1) above tries to highlight a connection between the interventional effects and target trials, but does not explain why emulating a target trial is desirable. a sentence or two explaining the merits of the trial emulation framework would help to strengthen this point.
3) can you briefly state what types of policy-relevant estimands moren-betancur et al. (2021) argue for? giving an example would help to clarify for the user what the type of intervention under study is.
4) under "statement of need," part of `medoutcon`'s functionality is mis-stated: the package implements asymptotically efficient causal machine learning-based estimators of both the natural and interventional direct/indirect effects. [per the `medoutcon` `README`](https://github.com/nhejazi/medoutcon#citation), please also cite Diaz et al. (2020, DOI: 10.1093/biomet/asaa085), which developed the methodology implemented in `medoutcon`. similar details should be given about both the `mediation` and `medflex` packages, about which i'm less knowledgeable.
5) in describing alternative software packages (`mediation`, `medflex`, `medouton`), only certain disadvantages of those packages are stated (e.g., challenging identifiability conditions of the NDE/NIE or causal machine learning of interventional effects not explicitly mapped to a target trial), but no disadvantages of `medRCT` are stated. it is, of course, fine to critique the alternatives available, but any gaps in what `medRCT` provides should be stated upfront too. for example, the "Monte Carlo simulation-based g-computation" approach, at minimum, (a) requires many Monte Carlo replications to provide reliable inference, and (b) is highly prone to model misspecification, as the nuisance parameters can only be estimated via restrictive parametric modeling strategies.
