### 2025-05-26 1e3638b
add statement of need in readme
