+++ README.md
@@ -26,6 +26,68 @@ mediators. `medRCT` can accommodate any number of potentially correlated
 mediators, including mediators that are not of primary interest but that
 are intermediate (exposure-induced) mediator-outcome confounders.
 
+## Statement of need:
+
+Causal mediation analysis generally seeks to investigate the extent to
+which the causal effect of an exposure on an outcome is mediated through
+intermediate variables. Natural (in)direct effects (Robins and Greenland
+1992; Pearl 2001) were initially proposed as the estimands of interest
+in these analyses. Natural effects are defined based on cross-world
+counterfactuals (Robins and Richardson 2011) and their identifiability
+relies on a cross-world independence assumption. Given their reliance on
+cross-world counterfactuals, these effects have been criticized for not
+capturing the effects of interventions or policy measures that could be
+conducted in the real world (Naimi, Kaufman, and MacLehose 2014).
+Further, the independence assumption required can never be guaranteed,
+even in an experiment, and it renders the estimands unidentifiable in
+the common settings of exposure-induced mediator-outcome confounding and
+multiple mediators. However, in the context of multiple mediators,
+certain path-specific natural effects, also defined in terms of
+cross-world counterfactuals, can still be identified and may be of
+substantive interest (VanderWeele and Vansteelandt 2014).
+
+Interventional effects have been proposed as an alternative to address
+these limitations. Firstly, these effects can be shown to map to a
+hypothetical randomized trial that evaluates the impact of hypothetical
+interventions shifting the distribution of the mediators
+(Moreno-Betancur and Carlin 2018). Secondly, interventional effects
+remain identifiable in the presence of exposure-induced mediator-outcome
+confounding and multiple interrelated mediators of interest.
+
+The `medRCT` package implements the estimation of interventional effects
+that are defined explicitly as effects in a hypothetical randomized
+trial (the target trial) , as proposed by Moreno-Betancur et al. (2021).
+This assists with clarifying the research question and ensuring that the
+study findings are meaningful and relevant to policy and practice. In
+the target trial, the treatment strategies are specified to reflect
+hypothetical interventions targeting and thus shifting the joint
+mediator distribution. The `medRCT` package implements the estimation of
+interventional effects that correspond to effects of hypothetical
+interventions which:
+
+1.  shift the joint distribution of all mediators under exposure to that
+    under no exposure
+
+2.  shift the distribution of a specific mediator under exposure, given
+    confounders, to match the corresponding distribution under no
+    exposure, independent of and without considering flow-on effects on
+    other mediators,
+
+3.  shift the distribution of a specific mediator under exposure, given
+    confounders, to match the corresponding distribution under no
+    exposure, while considering flow-on effects on causally descendant
+    mediators.
+
+`medRCT` estimates these interventional effects using a Monte Carlo
+simulation-based g-computation approach. It should be noted that this
+method can be computationally intensive and is sensitive to model
+misspecification, as all nuisance parameters are estimated via
+restrictive parametric models.
+
+Researchers should consider using `medRCT` when their ultimate goal for
+conducting mediation analysis is to examine the effects of hypothetical
[…]
