+++ inst/references.bib
@@ -1,6 +1,6 @@
 @article{vanderweele2014effect,
   title={Effect decomposition in the presence of an exposure-induced mediator-outcome confounder},
-  author={VanderWeele, Tyler J and Vansteelandt, Stijn and Robins, James M},
+  author={Tyler J VanderWeele and Stijn Vansteelandt and James M Robins},
   journal={Epidemiology},
   volume={25},
   number={2},
@@ -13,7 +13,7 @@
 
 
 @article{Geneletti2007Identifying,
-  author = {Geneletti, Sara},
+  author = {Sara Geneletti},
   title = {Identifying direct and indirect effects in a non-counterfactual framework},
   journal = {Journal of the Royal Statistical Society: Series B (Statistical Methodology)},
   volume = {69},
@@ -42,7 +42,7 @@
 
 @article{hernan2016using,
   title={Using big data to emulate a target trial when a randomized trial is not available},
-  author={Hern{\'a}n, Miguel A and Robins, James M},
+  author={Miguel A Hern{\'a}n and James M Robins},
   journal={American Journal of Epidemiology},
   volume={183},
   number={8},
@@ -56,7 +56,7 @@
 
 
 @article{MorenoBetancur2018,
-  author={Moreno-Betancur, Margarita and Carlin, John B.},
+  author={Margarita Moreno-Betancur and John B. Carlin},
   title={Understanding interventional effects: A more natural approach to mediation analysis},
   journal={Epidemiology},
   year={2018},
@@ -70,7 +70,7 @@
 
 
 @article{Robins1992,
-  author={Robins, James M. and Greenland, Sander},
+  author={James M Robins and Sander Greenland},
   title={Identifiability and exchangeability for direct and indirect effects},
   journal={Epidemiology},
   year={1992},
@@ -84,7 +84,7 @@
 
 
 @article{Vansteelandt2017,
-  author={Vansteelandt, Stijn and Daniel, Rhian M.},
+  author={Stijn Vansteelandt and Rhian M. Daniel},
   title={Interventional effects for mediation analysis with multiple mediators},
   journal={Epidemiology},
   year={2017},
@@ -139,8 +139,7 @@
 }
 
 @article{hejazi2022medoutcon,
-  author = {Hejazi, Nima S and Rudolph, Kara E and D{\'\i}az,
-        Iv{\'a}n},
+  author = {Nima S Hejazi and Kara E Rudolph and Iv{\'a}n D{\'\i}az},
   title = {{medoutcon}: Nonparametric efficient causal mediation
         analysis with machine learning in {R}},
   year = {2022},
@@ -154,7 +153,7 @@
 
 
 @article{Dashti2022,
-  author = {Dashti, S. Ghazaleh and Simpson, Julie A. and Viallon, Vivian and Karahalios, Amalia and Moreno-Betancur, Margarita and Brasky, Theodore and Pan, Kathy and Rohan, Thomas E. and Shadyab, Aladdin H. and Thomson, Cynthia A. and Wild, Robert A. and Wassertheil-Smoller, Sylvia and Ho, Gloria Y. F. and Strickler, Howard D. and English, Dallas R. and Gunter, Marc J.},
+  author = {S. Ghazaleh Dashti and Julie A Simpson and Vivian Viallon and Amalia Karahalios and Margarita Moreno-Betancur and Theodore Brasky and Kathy Pan and Thomas E. Rohan and Aladdin H. Shadyab and Cynthia A. Thomson and Robert A. Wild and Sylvia Wassertheil-Smoller and Gloria Y. F. Ho and Howard D. Strickler and Dallas R. English and Marc J. Gunter},
[…]
