+++ inst/references.bib
@@ -1,6 +1,6 @@
 @article{vanderweele2014effect,
   title={Effect decomposition in the presence of an exposure-induced mediator-outcome confounder},
-  author={Tyler J VanderWeele and Stijn Vansteelandt and James M Robins},
+  author={Tyler J. VanderWeele and Stijn Vansteelandt and James M. Robins},
   journal={Epidemiology},
   volume={25},
   number={2},
@@ -27,7 +27,7 @@
 
 
 @article{Moreno2021Mediation,
-  author={Margarita Moreno-Betancur and Paul Moran and Denise Becker and George C Patton and John B Carlin},
+  author={Margarita Moreno-Betancur and Paul Moran and Denise Becker and George C. Patton and John B. Carlin},
   title={Mediation effects that emulate a target randomised trial: Simulation-based evaluation of ill-defined interventions on multiple mediators},
   journal={Statistical Methods in Medical Research},
   volume={30},
@@ -42,7 +42,7 @@
 
 @article{hernan2016using,
   title={Using big data to emulate a target trial when a randomized trial is not available},
-  author={Miguel A Hern{\'a}n and James M Robins},
+  author={Miguel A Hern{\'a}n and James M. Robins},
   journal={American Journal of Epidemiology},
   volume={183},
   number={8},
@@ -70,7 +70,7 @@
 
 
 @article{Robins1992,
-  author={James M Robins and Sander Greenland},
+  author={James M. Robins and Sander Greenland},
   title={Identifiability and exchangeability for direct and indirect effects},
   journal={Epidemiology},
   year={1992},
@@ -139,7 +139,7 @@
 }
 
 @article{hejazi2022medoutcon,
-  author = {Nima S Hejazi and Kara E Rudolph and Iv{\'a}n D{\'\i}az},
+  author = {Nima S. Hejazi and Kara E. Rudolph and Iv{\'a}n D{\'\i}az},
   title = {{medoutcon}: Nonparametric efficient causal mediation
         analysis with machine learning in {R}},
   year = {2022},
@@ -153,7 +153,7 @@
 
 
 @article{Dashti2022,
-  author = {S. Ghazaleh Dashti and Julie A Simpson and Vivian Viallon and Amalia Karahalios and Margarita Moreno-Betancur and Theodore Brasky and Kathy Pan and Thomas E. Rohan and Aladdin H. Shadyab and Cynthia A. Thomson and Robert A. Wild and Sylvia Wassertheil-Smoller and Gloria Y. F. Ho and Howard D. Strickler and Dallas R. English and Marc J. Gunter},
+  author = {S. Ghazaleh Dashti and Julie A. Simpson and Vivian Viallon and Amalia Karahalios and Margarita Moreno-Betancur and Theodore Brasky and Kathy Pan and Thomas E. Rohan and Aladdin H. Shadyab and Cynthia A. Thomson and Robert A. Wild and Sylvia Wassertheil-Smoller and Gloria Y. F. Ho and Howard D. Strickler and Dallas R. English and Marc J. Gunter},
   title = {Adiposity and breast, endometrial, and colorectal cancer risk in postmenopausal women: Quantification of the mediating effects of leptin, C-reactive protein, fasting insulin, and estradiol},
   journal = {Cancer Medicine},
   volume = {11},
@@ -211,7 +211,7 @@ url = {https://dl.acm.org/doi/pdf/10.5555/2074022.2074073}
 
 
 @incollection{Robins2011,
-    author = {James M Robins and Thomas S Richardson},
+    author = {James M. Robins and Thomas S. Richardson},
     title = {Alternative graphical causal models and the identification of direct effects},
[…]
