+++ README.md
@@ -39,12 +39,15 @@ cross-world counterfactuals, these effects have been criticized for not
 capturing the effects of interventions or policy measures that could be
 conducted in the real world (Naimi, Kaufman, and MacLehose 2014).
 Further, the independence assumption required can never be guaranteed,
-even in an experiment, and it renders the estimands unidentifiable in
-the common settings of exposure-induced mediator-outcome confounding and
-multiple mediators. However, in the context of multiple mediators,
-certain path-specific natural effects, also defined in terms of
-cross-world counterfactuals, can still be identified and may be of
-substantive interest (VanderWeele and Vansteelandt 2014).
+even in an experiment (Robins and Richardson 2011; Didelez, Dawid, and
+Geneletti 2006), and it renders the estimands unidentifiable in the
+common settings of exposure-induced mediator-outcome confounding and
+multiple mediators (Avin, Shpitser, and Pearl 2005; VanderWeele,
+Vansteelandt, and Robins 2014; Vansteelandt and VanderWeele 2012).
+However, in the context of multiple mediators, certain path-specific
+natural effects, also defined in terms of cross-world counterfactuals,
+can still be identified and may be of substantive interest (VanderWeele
+and Vansteelandt 2014).
 
 Interventional effects have been proposed as an alternative to address
 these limitations. Firstly, these effects can be shown to map to a
@@ -66,7 +69,7 @@ interventional effects that correspond to effects of hypothetical
 interventions which:
 
 1.  shift the joint distribution of all mediators under exposure to that
-    under no exposure
+    under no exposure,
 
 2.  shift the distribution of a specific mediator under exposure, given
     confounders, to match the corresponding distribution under no
@@ -239,6 +242,26 @@ For work involving the `medRCT` R package, please cite the following:
 <div id="refs" class="references csl-bib-body hanging-indent"
 entry-spacing="0">
 
+<div id="ref-Avin2005" class="csl-entry">
+
+Avin, Chen, Ilya Shpitser, and Judea Pearl. 2005. “Identifiability of
+Path-Specific Effects.” In *Proceedings of the 19th International Joint
+Conference on Artificial Intelligence*, 357–63. IJCAI’05. San Francisco,
+CA, USA: Morgan Kaufmann Publishers Inc.
+<https://doi.org/10.5555/1642293.1642350>.
+
+</div>
+
+<div id="ref-Didelez2006" class="csl-entry">
+
+Didelez, Vanessa, Philip Dawid, and Sara Geneletti. 2006. “Direct and
+Indirect Effects of Sequential Treatments.” In *Proceedings of the
+Twenty-Second Conference on Uncertainty in Artificial Intelligence*,
+138–46. UAI’06. Arlington, Virginia, USA: AUAI Press.
+<https://doi.org/10.5555/3020419.3020437>.
+
+</div>
+
 <div id="ref-Goldfeld2023" class="csl-entry">
 
 Goldfeld, Sharon, Margarita Moreno-Betancur, Sarah Gray, Shuaijun Guo,
@@ -251,8 +274,8 @@ Health and Preschool Attendance.” *Pediatrics* 151 (5): e2022057101.
 
 <div id="ref-hernan2016using" class="csl-entry">
 
-Hernán, Miguel A, and James M Robins. 2016. “Using Big Data to Emulate a
[…]

[…]
