+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   intro: intro.html
-last_built: 2025-02-25T06:10Z
+last_built: 2025-05-27T10:11Z
 urls:
   reference: https://t0ngchen.github.io/medRCT/reference
   article: https://t0ngchen.github.io/medRCT/articles
