### 2025-05-27 #8063 @T0ngChen
Hi @nhejazi, thanks so much for your detailed comments, which have helped us improve the clarity and completeness of the documentation and manuscript! We appreciate your insights and suggestions. Please find the detailed changes below.

**documentation -- statement of need**

> https://github.com/T0ngChen/medRCT/issues/23 details a few changes that could be made to help tick off the "statement of need"

As suggested, we have added a “Statement of Need” section to the README. Further details are provided in [issue #23](https://github.com/T0ngChen/medRCT/issues/23).

**software paper**

> please clarify what the "certain identifiability challenges associated with traditional estimands, known as the natural effects" are (also Pearl (2001) should be cited alongside Robins and Greenland (1992) for the NDE/NIE). explaining this in detail is beyond the scope of a software paper but this statement is too vague for the inexperienced reader to appreciate what exactly the challenges are.

Thank you for highlighting this important point. We have added a paragraph to elaborate on these challenges, and have also cited Pearl (2001) alongside Robins and Greenland (1992) as recommended (page 1 lines 15-24):

"The modern causal inference literature initially defined the natural (in)direct effects (Pearl, 2001; Robins & Greenland, 1992) as the estimands of interest in causal mediation analysis. Grounded in the potential outcomes framework, these effects are defined based on cross-world counterfactuals (Robins & Richardson, 2011), and their identifiability relies on a cross-world independence assumption. This assumption, however, can never be guaranteed to hold, not even with an experiment (Didelez et al., 2006; Robins & Richardson, 2011). Further, this assumption is not met in the presence of a mediator-outcome confounder that is itself affected by the exposure (Avin et al., 2005; VanderWeele et al., 2014; Vansteelandt & VanderWeele, 2012), and more generally, in the presence of multiple mediators of interest (VanderWeele & Vansteelandt, 2014)."

> the second half of the paragraph containing the excerpt in (1) above tries to highlight a connection between the interventional effects and target trials, but does not explain why emulating a target trial is desirable. a sentence or two explaining the merits of the trial emulation framework would help to strengthen this point.

Thanks for pointing this out. In the revised version, we have clarified why the trial emulation framework is desirable (page 1, lines 25-32):

"Interventional effects (Geneletti, 2007; VanderWeele et al., 2014) have been proposed as an alternative to address these limitations. In particular, these effects have been shown to implicitly emulate target trials (Hernán & Robins, 2016) that assess the impact of hypothetical interventions shifting the distribution of the mediators (Moreno-Betancur & Carlin, 2018). It has thus been proposed that interventional effects be explicitly defined by mapping to a hypothetical randomised trial (a target trial) that assesses the hypothetical interventions of interest (Moreno-Betancur et al., 2021). Specifying the target trial clarifies the causal question and makes the interventional effects directly policy-relevant and practically meaningful."

> can you briefly state what types of policy-relevant estimands moren-betancur et al. (2021) argue for? giving an example would help to clarify for the user what the type of intervention under study is.

We have now briefly summarised the types of policy-relevant estimands discussed in Moren-Betancur et al. (2021), including an example to help clarify the type of intervention under study (page 1, lines 33-43):

"Moreno-Betancur et al. (2021) proposed three interventional effects, with each one examining a distinct policy-relevant question about how hypothetical interventions that would shift mediator distributions individually, together, or in sequence might impact the outcome. For example, they examined the extent to which the increased risk of financial hardship among adolescent self-harmers could be countered by a hypothetical intervention that would shift the distribution of weekly cannabis use among adolescent self-harmers to the level of those who did not self-harm, either treating other mediators as independent or allowing for flow-on effects on its causal descendants, such as education and employment. They also considered an interventional effect capturing the effect of an intervention that shifts the joint distribution of all mediators (in their example, depression or anxiety, cannabis use, education, and employment status) in the exposed to the levels in the unexposed group."

> under "statement of need," part of medoutcon's functionality is mis-stated: the package implements asymptotically efficient causal machine learning-based estimators of both the natural and interventional direct/indirect effects. [per the medoutcon README](https://github.com/nhejazi/medoutcon#citation), please also cite Diaz et al. (2020, DOI: 10.1093/biomet/asaa085), which developed the methodology implemented in medoutcon. similar details should be given about both the mediation and medflex packages, about which i'm less knowledgeable.

Thank you for pointing out the inaccuracies and lack of details. We have provided details for the existing software (page 2, lines 51-63):
"Several R (R Core Team, 2025) software packages are available for causal mediation analysis. The mediation package (Tingley et al., 2014) implements the estimation of natural effects using a generic approach based on Monte-Carlo integration methods as described by Imai et al. (2010). The package can also estimate path-specific effects when there are multiple mediators, and allows researchers to conduct sensitivity analyses to evaluate the robustness of their results to potential unmeasured confounding. The medflex package (Steen et al., 2017) implements the methods proposed by Vansteelandt et al. (2012), which directly model the natural effects based on a class of mean models for nested counterfactuals. The medoutcon package (Hejazi et al., 2022) implements asymptotically efficient causal machine learning-based estimators of both the natural and interventional direct and indirect effects (Dıá z et al., 2020). However, the interventional effects considered are not defined explicitly in terms of a target trial examining policy-relevant interventions, which may limit their direct practical relevance in informing decision-making"

> in describing alternative software packages (mediation, medflex, medouton), only certain disadvantages of those packages are stated (e.g., challenging identifiability conditions of the NDE/NIE or causal machine learning of interventional effects not explicitly mapped to a target trial), but no disadvantages of medRCT are stated. it is, of course, fine to critique the alternatives available, but any gaps in what medRCT provides should be stated upfront too. for example, the "Monte Carlo simulation-based g-computation" approach, at minimum, (a) requires many Monte Carlo replications to provide reliable inference, and (b) is highly prone to model misspecification, as the nuisance parameters can only be estimated via restrictive parametric modeling strategies.

We appreciate your suggestion to discuss the limitations of our package. We have added the following statement to page 2, lines 65-71:

''However, as medRCT uses a Monte Carlo simulation-based g-computation approach, it requires a large number of replications to produce stable and reliable inference, which can be computationally intensive. Additionally, the approach is sensitive to model misspecification, as all nuisance parameters can only be estimated via restrictive parametric modelling. Furthermore, the method is not suited for settings with high-dimensional mediators or intermediate confounders, where causal machine learning-based estimators are required (Chen et al., 2025; Dıá z et al., 2020; Liu et al., 2024; Rudolph et al., 2024).''
