+++ paper/paper.md
@@ -38,9 +38,9 @@ The modern causal inference literature initially defined the natural (in)direct 
 Interventional effects [@Geneletti2007Identifying; @vanderweele2014effect] have been proposed as an alternative to address these limitations. In particular, these effects have been shown to implicitly emulate target trials [@hernan2016using] that assess the impact of hypothetical interventions shifting the distribution of the mediators [@MorenoBetancur2018]. It has thus been proposed that interventional effects be explicitly defined by mapping to a hypothetical randomised trial (a target trial) that assesses the hypothetical interventions of interest [@Moreno2021Mediation]. Specifying the target trial clarifies the causal question and makes the interventional effects directly policy-relevant and practically meaningful.
 
 
-@Moreno2021Mediation proposed three interventional effects, with each one examining a distinct policy-relevant question about how hypothetical interventions that would shift mediator distributions individually, together, or in sequence might impact the outcome. For example, they examined the extent to which the increased risk of financial hardship among adolescent self-harmers could be countered by a hypothetical intervention that would shift the distribution of weekly cannabis use among adolescent self-harmers to the level of those who did not self-harm, either treating other mediators as independent or allowing for flow-on effects on its causal descendants, such as education and employment. They also considered an interventional effect capturing the effect of an intervention that shifts the joint distribution of all mediators (in their example, depression or anxiety, cannabis use, education, and employment status) in the exposed to the levels in the unexposed group.
+@Moreno2021Mediation proposed three interventional effects, with each one examining a distinct policy-relevant question about how hypothetical interventions that would shift mediator distributions individually, together, or in sequence might impact the outcome. For example, they examined the extent to which the increased risk of financial hardship in mid-adulthood (outcome) among adolescent self-harmers (exposed group) could be countered by a hypothetical intervention that would shift the distribution of weekly cannabis use in young adulthood (mediator) among adolescent self-harmers to the levels of those who did not self-harm (unexposed group), either treating other mediators as independent or allowing for flow-on effects on its causal descendants, such as education and employment. They also considered an interventional effect capturing the effect of an intervention that shifts the joint distribution of all mediators (in their example, depression or anxiety, cannabis use in young adulthood, education, and employment status) among adolescent self-harmers to the levels in those who did not self-harm.
 
-The `medRCT` package provides a user-friendly interface for estimating policy-relevant interventional effects as defined in @Moreno2021Mediation using a Monte Carlo simulation-based g-computation approach. Key features of `medRCT` include support for multiple mediators, intermediate confounders (exposure-induced mediator-outcome confounders), and a Shiny application [@chang2024shiny] for comprehensive model assessment. Detailed methods and estimands implemented in `medRCT` are described in @Moreno2021Mediation.
+The `medRCT` package provides a user-friendly interface for estimating policy-relevant interventional effects as defined in @Moreno2021Mediation using a Monte Carlo simulation-based g-computation approach. Key features of `medRCT` include support for multiple mediators, intermediate confounders (exposure-induced mediator-outcome confounders), and a Shiny application [@chang2024shiny] for comprehensive model assessment. Detailed definition of the mediation estimands that can be estimated using the `medRCT` package, as well as the target trial to which they map, their identifiability and estimation procedures are described in @Moreno2021Mediation.
 
 # Statement of need
 
@@ -50,7 +50,7 @@ Several R [@R] software packages are available for causal mediation analysis. Th
[…]
