### 2025-06-11 e01a714
document effect_measure argument
