### 2025-06-11 887b774
add effect_measure argument and handle warnings/errors with tryCatch
