+++ R/est_fun.R
@@ -31,32 +31,36 @@
 #'    specific mediator (\code{k}) in the exposed to match the corresponding distribution in the unexposed while accounting for the flow-on
 #'    effects on causally descendant mediators.
 #' }
+#' @param effect_measure A \code{character} string specifying the effect measure to calculate.
+#'   For a continuous outcome, only \code{"Diff"} (difference in means) is allowed.
+#'   For a binary outcome, must be either \code{"RD"} (risk difference) or \code{"RR"} (risk ratio).
+#'   If not specified, defaults to \code{"Diff"} for continuous outcomes and \code{"RD"} for binary outcomes.
+#'   Only one effect measure can be specified at a time.
 #' @param mcsim An \code{integer} specifying the number of Monte Carlo simulations to perform.
 #'
 #' @importFrom stats as.formula binomial glm predict rbinom rnorm df.residual
 #' @importFrom data.table as.data.table ":="
 #'
 #' @keywords internal
-medRCT.fun <- function(dat,
-                       ind = 1:nrow(dat),
-                       first = first,
-                       K = K,
-                       fam_type = fam_type,
-                       mediators = mediators,
-                       interactions_XC = interactions_XC,
-                       intervention_type = intervention_type,
-                       mcsim) {
+medRCT.fun <- function(
+  dat,
+  ind = 1:nrow(dat),
+  first = first,
+  K = K,
+  fam_type = fam_type,
+  mediators = mediators,
+  interactions_XC = interactions_XC,
+  intervention_type = intervention_type,
+  effect_measure,
+  mcsim
+) {
   # Take bootstrap sample
   data <- dat[ind, ]
-
-  # Set flag to capture bootstrap samples to reject
-  flag <- FALSE
 
   # Replicate dataset for simulations
   dat2 <- data.table::as.data.table(data)
 
-  dat2[, 1:(2 + K) := lapply(.SD,
-                             function(x) NA_integer_), .SDcols = 1:(2 + K)]
+  dat2[, 1:(2 + K) := lapply(.SD, function(x) NA_integer_), .SDcols = 1:(2 + K)]
 
   dat2 <- zoo::coredata(dat2)[rep(seq(nrow(dat2)), mcsim), ]
   n <- nrow(dat2)
@@ -64,37 +68,60 @@ medRCT.fun <- function(dat,
   # identify the exposure levels
   data$X = as.factor(data$X)
   exposure_level = sort(unique(as.numeric(dat$X)))
-  lnzero = exposure_level[exposure_level!=0]
+  lnzero = exposure_level[exposure_level != 0]
 
   # ESTIMATE DISTRIBUTIONS
   # Joint of M1 to MK under X=0 and X!=0 ...
 
   for (k in 1:K) {
-    dat2 = joint_dist(k = k, K = K, data = data, dat2 = dat2, mediators = mediators,
-                      fam_type = fam_type, interactions_XC = interactions_XC,
-                      exposure_level = exposure_level, n = n)
+    dat2 = joint_dist(
+      k = k,
+      K = K,
+      data = data,
+      dat2 = dat2,
+      mediators = mediators,
+      fam_type = fam_type,
+      interactions_XC = interactions_XC,
+      exposure_level = exposure_level,
+      n = n
+    )
   }
 
   # Estimating the target quantities
   # Marginals under X=0
   for (k in first:K) {
     dat2 <- marg_dist(
-      k = k, first = first, K = K, data = data, dat2 = dat2, mediators = mediators,
-      fam_type = fam_type, interactions_XC = interactions_XC, n = n
+      k = k,
+      first = first,
+      K = K,
+      data = data,
+      dat2 = dat2,
+      mediators = mediators,
[…]
