### 2025-06-11 3a82dd9
calculate results according to effect_measure
