+++ R/est_proc.R
@@ -1,4 +1,3 @@
-
 #' Estimation of Joint Distribution of Mediators and Intermediate Confounders under
 #' Exposed and Unexposed
 #'
@@ -22,27 +21,46 @@
 #' @importFrom stats coef as.formula
 #'
 #' @keywords internal
-joint_dist <- function(k, K, data, dat2, fam_type, mediators,
-                       interactions_XC, exposure_level, n) {
+joint_dist <- function(
+  k,
+  K,
+  data,
+  dat2,
+  fam_type,
+  mediators,
+  interactions_XC,
+  exposure_level,
+  n
+) {
   # Fit the model
   fit <- glm(
-    stats::as.formula(gen_formula(k = k,
-                                  interactions_XC = interactions_XC,
-                                  include_all = TRUE)),
+    stats::as.formula(gen_formula(
+      k = k,
+      interactions_XC = interactions_XC,
+      include_all = TRUE
+    )),
     data = data,
     family = fam_type[[k]]
   )
 
   # Check convergence and coefficients
-  if(!fit$converged){
-    warning(paste0("Model did not converge when using variable ", mediators[k], " as the response"))
+  if (!fit$converged) {
+    warning(paste0(
+      "Model did not converge when using variable ",
+      mediators[k],
+      " as the response"
+    ))
   }
-  if(any(is.na(fit$coefficients))){
+  if (any(is.na(fit$coefficients))) {
     na_coefs <- names(stats::coef(fit))[is.na(stats::coef(fit))]
-    warning(paste0("The following coefficients were NA: ", paste(na_coefs, collapse = ", "),
-                   "when using variable ", mediators[k], " as the response"))
+    warning(paste0(
+      "The following coefficients were NA: ",
+      paste(na_coefs, collapse = ", "),
+      "when using variable ",
+      mediators[k],
+      " as the response"
+    ))
   }
-
 
   # Loop through exposure levels
   for (a in exposure_level) {
@@ -68,7 +86,6 @@ joint_dist <- function(k, K, data, dat2, fam_type, mediators,
 }
 
 
-
 #' Estimation of Marginal Distributions of Mediators under Unexposed
 #'
 #' @param k An integer specifying the index of the current mediator being processed.
@@ -90,25 +107,44 @@ joint_dist <- function(k, K, data, dat2, fam_type, mediators,
 #' @importFrom stats coef as.formula
 #'
 #' @keywords internal
-marg_dist <- function(k, first, K, data, dat2, fam_type, mediators,
-                      interactions_XC, n) {
+marg_dist <- function(
+  k,
+  first,
+  K,
+  data,
+  dat2,
+  fam_type,
+  mediators,
+  interactions_XC,
+  n
+) {
   # Fit the model
   fit <- glm(
-    stats::as.formula(gen_formula(k = k,
-                                  interactions_XC = interactions_XC,
-                                  marginal = TRUE)),
+    stats::as.formula(gen_formula(
+      k = k,
+      interactions_XC = interactions_XC,
+      marginal = TRUE
+    )),
     data = data,
     family = fam_type[[k]]
   )
 
-  # Initialize a flag for convergence issues
-  if(!fit$converged){
-    warning(paste0("Model did not converge when using variable ", mediators[k], " as the response"))
+  if (!fit$converged) {
+    warning(paste0(
+      "Model did not converge when using variable ",
+      mediators[k],
+      " as the response"
+    ))
   }
-  if(any(is.na(fit$coefficients))){
+  if (any(is.na(fit$coefficients))) {
     na_coefs <- names(stats::coef(fit))[is.na(stats::coef(fit))]
[…]
