### 2025-06-11 771198c
add effect_measure argument, disregard failed bootstraps, fix factor mediator bug in main function
