+++ R/medRCT.R
@@ -37,6 +37,11 @@ utils::globalVariables(".SD")
 #'    specific mediator (\code{k}) in the exposed, given baseline confounders, to match the corresponding distribution in the unexposed while accounting for the flow-on
 #'    effects on causally descendant mediators.
 #' }
+#' @param effect_measure A \code{character} string specifying the effect measure to calculate.
+#'   For a continuous outcome, only \code{"Diff"} (difference in means) is allowed.
+#'   For a binary outcome, must be either \code{"RD"} (risk difference) or \code{"RR"} (risk ratio).
+#'   If not specified, defaults to \code{"Diff"} for continuous outcomes and \code{"RD"} for binary outcomes.
+#'   Only one effect measure can be specified at a time.
 #' @param mcsim An \code{integer} specifying the number of Monte Carlo simulations to perform. The default is 200.
 #' It is recommended to run analysis with no fewer than 200 Monte Carlo simulations.
 #' @param bootstrap A \code{logical} value indicating whether bootstrapping should be performed. If \code{TRUE}
@@ -59,6 +64,9 @@ utils::globalVariables(".SD")
 #' If issues with model fitting are detected, users are encouraged to adjust the exposure-confounder interaction term as needed.
 #' However, \strong{mediators or confounders must not be selected based on model fitting results.}
 #'
+#' The function accepts binary mediators and intermediate confounders coded as 0/1, TRUE/FALSE, or any two unique values (numeric, character, or factor).
+#' Internally, the lower value will always be treated as 0 and the higher value as 1, regardless of the original coding.
+#'
 #' @export
 #'
 #' @examples
@@ -80,6 +88,7 @@ utils::globalVariables(".SD")
 #'   confounders = confounders,
 #'   interactions_XC = "all",
 #'   intervention_type = "all",
+#'   effect_measure = "RD",
 #'   bootstrap = TRUE,
 #'   boot_args = list(R = 100, stype = "i", ci.type = "norm"),
 #'   mcsim = 100
@@ -87,36 +96,44 @@ utils::globalVariables(".SD")
 #' # Summarise the results
 #' summary(med_res)
 #' }
-medRCT <- function(dat,
-                   exposure,
-                   outcome,
-                   mediators,
-                   intermediate_confs,
-                   confounders,
-                   interactions_XC = "all",
-                   intervention_type = c("all", "shift_all", "shift_k", "shift_k_order"),
-                   mcsim = 200,
-                   bootstrap = TRUE,
-                   boot_args = list(R = 100, stype = "i", ci.type = "norm"),
-                   ...) {
+medRCT <- function(
+  dat,
+  exposure,
+  outcome,
+  mediators,
+  intermediate_confs,
+  confounders,
+  interactions_XC = "all",
+  intervention_type = c("all", "shift_all", "shift_k", "shift_k_order"),
+  effect_measure = NULL,
+  mcsim = 200,
+  bootstrap = TRUE,
+  boot_args = list(R = 100, stype = "i", ci.type = "norm"),
+  ...
+) {
   # match intervention type
-  intervention_type = sapply(intervention_type, function(arg)
-    match.arg(
-      arg,
-      choices = c("all", "shift_all", "shift_k", "shift_k_order")
-    ))
+  choices <- c("all", "shift_all", "shift_k", "shift_k_order")
+  idx <- match(intervention_type, choices)
+  intervention_type <- choices[idx]
 
[…]
