### 2025-06-20 T0ngChen/medRCT#25 @fabian-s
[joss] final tidying up
fix typos / language please:

- mixed "Monte-Carlo" (hyphenated) vs "Monte Carlo"  - pick one
- subject–verb clash & wrong verb: “detailed definition of the mediation estimands … are described” should be "definition[s] ... are [given]"
- "has been extensively documented on its website” should be "is extensively documented at its website"

see https://github.com/openjournals/joss-reviews/issues/8063
