### 2025-06-20 256c8ac
[joss] final tidying up -- fix typos / language
