+++ paper/paper.md
@@ -40,11 +40,11 @@ Interventional effects [@Geneletti2007Identifying; @vanderweele2014effect] have 
 
 @Moreno2021Mediation proposed three interventional effects, with each one examining a distinct policy-relevant question about how hypothetical interventions that would shift mediator distributions individually, together, or in sequence might impact the outcome. For example, they examined the extent to which the increased risk of financial hardship in mid-adulthood (outcome) among adolescent self-harmers (exposed group) could be countered by a hypothetical intervention that would shift the distribution of weekly cannabis use in young adulthood (mediator) among adolescent self-harmers to the levels of those who did not self-harm (unexposed group), either treating other mediators as independent or allowing for flow-on effects on its causal descendants, such as education and employment. They also considered an interventional effect capturing the effect of an intervention that shifts the joint distribution of all mediators (in their example, depression or anxiety, cannabis use in young adulthood, education, and employment status) among adolescent self-harmers to the levels in those who did not self-harm.
 
-The `medRCT` package provides a user-friendly interface for estimating policy-relevant interventional effects as defined in @Moreno2021Mediation using a Monte Carlo simulation-based g-computation approach. Key features of `medRCT` include support for multiple mediators, intermediate confounders (exposure-induced mediator-outcome confounders), and a Shiny application [@chang2024shiny] for comprehensive model assessment. Detailed definition of the mediation estimands that can be estimated using the `medRCT` package, as well as the target trial to which they map, their identifiability and estimation procedures are described in @Moreno2021Mediation.
+The `medRCT` package provides a user-friendly interface for estimating policy-relevant interventional effects as defined in @Moreno2021Mediation using a Monte Carlo simulation-based g-computation approach. Key features of `medRCT` include support for multiple mediators, intermediate confounders (exposure-induced mediator-outcome confounders), and a Shiny application [@chang2024shiny] for comprehensive model assessment. Detailed definitions of the mediation estimands that can be estimated using the `medRCT` package, as well as the target trial to which they map, their identifiability and estimation procedures are given in @Moreno2021Mediation.
 
 # Statement of need
 
-Several R [@R] software packages are available for causal mediation analysis. The `mediation` package [@Tingley2014mediation] implements the estimation of natural effects using a generic approach based on Monte-Carlo integration methods as described by @Imai2010. The package can also estimate path-specific effects when there are multiple mediators, and allows researchers to conduct sensitivity analyses to evaluate the robustness of their results to potential unmeasured confounding. The `medflex` package [@Steen2017medflex] implements the methods proposed by @Vansteelandt2012Imputation, which directly model the natural effects based on a class of mean models for nested counterfactuals. The `medoutcon` package [@hejazi2022medoutcon] implements asymptotically efficient causal machine learning-based estimators of both the natural and interventional direct and indirect effects [@Díaz2020]. However, the interventional effects considered are not defined explicitly in terms of a target trial examining policy-relevant interventions, which may limit their direct practical relevance in informing decision-making. 
[…]
