+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   intro: intro.html
-last_built: 2025-05-27T11:23Z
+last_built: 2025-06-20T09:39Z
 urls:
   reference: https://t0ngchen.github.io/medRCT/reference
   article: https://t0ngchen.github.io/medRCT/articles
