### 2025-06-20 #8063 @T0ngChen
@nhejazi Thank you so much for reviewing the changes and for your helpful comments throughout the process.

@fabian-s Thanks so much for the guidance and for overseeing the review process. I’ve addressed the remaining typos in this [commit](https://github.com/T0ngChen/medRCT/commit/256c8ac) and completed all additional author tasks:

1. Double checked authors and affiliations
2. Made a release on GitHub ([v.0.1.0](https://github.com/T0ngChen/medRCT/releases/tag/v.0.1.0))
3. Archived the release on Zenodo with doi: [10.5281/zenodo.15704560](https://doi.org/10.5281/zenodo.15704560)
4. Made sure that the title and author list in the archive match those in the JOSS paper.
5. Confirmed that the MIT license for the archive matches the software license.
