### 2025-06-20 #8063 @fabian-s
@T0ngChen sorry,  I should have done the reference check FIRST -- my mistake, apologies...!

please add the DOI the bot suggests for the R manuals as well and correct all the broken ones as well,  then check them (`@editorialbot check references`) and update the deposits/releases accordingly...
