+++ inst/references.bib
@@ -206,7 +206,7 @@ pages = {411–420},
 numpages = {10},
 series = {UAI'01},
 doi={10.5555/2074022.2074073},
-url = {https://dl.acm.org/doi/pdf/10.5555/2074022.2074073}
+url = {https://dl.acm.org/doi/10.5555/2074022.2074073}
 }
 
 
@@ -230,7 +230,7 @@ booktitle = {Proceedings of the Twenty-Second Conference on Uncertainty in Artif
 pages = {138–146},
 numpages = {9},
 doi = {10.5555/3020419.3020437},
-url = {https://dl.acm.org/doi/pdf/10.5555/3020419.3020437},
+url = {https://dl.acm.org/doi/10.5555/3020419.3020437},
 series = {UAI'06}
 }
 
@@ -257,7 +257,7 @@ booktitle = {Proceedings of the 19th International Joint Conference on Artificia
 pages = {357–363},
 numpages = {7},
 doi = {10.5555/1642293.1642350},
-url = {https://dl.acm.org/doi/pdf/10.5555/1642293.1642350},
+url = {https://dl.acm.org/doi/10.5555/1642293.1642350},
 series = {IJCAI'05}
 }
 
@@ -354,7 +354,8 @@ pages = {131--158},
 volume = {1},
 number = {1},
 journal = {Epidemiologic Methods},
-doi = {doi:10.1515/2161-962X.1014},
+doi = {10.1515/2161-962X.1014},
+URL = {https://doi.org/10.1515/2161-962X.1014}
 year = {2012},
 }
 
@@ -379,5 +380,6 @@ year = {2012},
   organization = {R Foundation for Statistical Computing},
   title = {{R}: A Language and Environment for Statistical Computing},
   url = {https://R-project.org/},
+  doi = {10.32614/r.manuals}
   year = {2025}
 }
