+++ inst/references.bib
@@ -355,7 +355,7 @@ volume = {1},
 number = {1},
 journal = {Epidemiologic Methods},
 doi = {10.1515/2161-962X.1014},
-URL = {https://doi.org/10.1515/2161-962X.1014}
+URL = {https://doi.org/10.1515/2161-962X.1014},
 year = {2012},
 }
 
@@ -380,6 +380,6 @@ year = {2012},
   organization = {R Foundation for Statistical Computing},
   title = {{R}: A Language and Environment for Statistical Computing},
   url = {https://R-project.org/},
-  doi = {10.32614/r.manuals}
+  doi = {10.32614/r.manuals},
   year = {2025}
 }
