+++ inst/references.bib
@@ -205,7 +205,6 @@ booktitle = {Proceedings of the 17th Conference on Uncertainty in Artificial Int
 pages = {411–420},
 numpages = {10},
 series = {UAI'01},
-doi={10.5555/2074022.2074073},
 url = {https://dl.acm.org/doi/10.5555/2074022.2074073}
 }
 
@@ -229,7 +228,6 @@ address = {Arlington, Virginia, USA},
 booktitle = {Proceedings of the Twenty-Second Conference on Uncertainty in Artificial Intelligence},
 pages = {138–146},
 numpages = {9},
-doi = {10.5555/3020419.3020437},
 url = {https://dl.acm.org/doi/10.5555/3020419.3020437},
 series = {UAI'06}
 }
@@ -256,7 +254,6 @@ address = {San Francisco, CA, USA},
 booktitle = {Proceedings of the 19th International Joint Conference on Artificial Intelligence},
 pages = {357–363},
 numpages = {7},
-doi = {10.5555/1642293.1642350},
 url = {https://dl.acm.org/doi/10.5555/1642293.1642350},
 series = {IJCAI'05}
 }
