### 2025-06-21 #8063 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
