### 2025-06-27 af66e04
return log RR
