### 2025-06-27 59a3c79
[joss] fix encoding in the metadata/title
