### 2025-06-28 a9a4807
add joss paper & markdown badge
