+++ README.Rmd
@@ -14,12 +14,13 @@ knitr::opts_chunk$set(
 )
 ```
 
-# medRCT
+# medRCT <a href="https://t0ngchen.github.io/medRCT"><img src="man/figures/logo.png" align="right" height="138" alt="medRCT website" /></a>
 
 <!-- badges: start -->
 [![MIT license](http://img.shields.io/badge/license-MIT-brightgreen.svg)](http://opensource.org/licenses/MIT)
 [![R-CMD-check](https://github.com/T0ngChen/medRCT/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/T0ngChen/medRCT/actions/workflows/R-CMD-check.yaml)
 [![Codecov test coverage](https://codecov.io/gh/T0ngChen/medRCT/graph/badge.svg)](https://app.codecov.io/gh/T0ngChen/medRCT)
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.08063/status.svg)](https://doi.org/10.21105/joss.08063)
 <!-- badges: end -->
 
 > Causal mediation analysis estimating interventional effects mapped to a target trial
@@ -102,13 +103,18 @@ For detailed guidance on using the package to handle more complex scenarios, ple
 
 For work involving the `medRCT` R package, please cite the following:
 
-    @software{Chen2024medRCT,
-       author = {Tong Chen and Margarita Moreno-Betancur and S. Ghazaleh Dashti},
-       title = {`r gsub("\n", " ", paste0("medRCT: ", utils::packageDescription("medRCT", fields = "Title")))`},
-       year  = {`r format(Sys.Date(), "%Y")`},
-       url = {https://t0ngchen.github.io/medRCT/},
-       note = {R package version `r utils::packageDescription("medRCT", fields = "Version")`}
-       }
+    @article{Chen2025medRCT,
+      author = {Tong Chen and S. Ghazaleh Dashti and Margarita Moreno-Betancur},
+      title = {{medRCT}: Causal mediation analysis estimating interventional effects mapped to a target trial in {R}},
+      year = {2025},
+      doi = {10.21105/joss.08063},
+      url = {https://doi.org/10.21105/joss.08063},
+      journal = {Journal of Open Source Software},
+      volume = {10},
+      number = {110},
+      pages = {8063},
+      publisher = {The Open Journal}
+    }
     @article{Moreno2021Mediation,
        author={Margarita Moreno-Betancur and Paul Moran and Denise Becker and George C Patton and John B Carlin},
        title={Mediation effects that emulate a target randomised trial: Simulation-based evaluation of ill-defined interventions on multiple mediators},
@@ -120,7 +126,6 @@ For work involving the `medRCT` R package, please cite the following:
        URL={https://doi.org/10.1177/0962280221998409},
        doi={10.1177/0962280221998409},
        publisher={SAGE Publications Ltd}
-       }    
-
+    }
 
 ## References
